Move App into components
